+++ python/sglang/test/test_utils.py
@@ -1564,7 +1564,7 @@ def run_bench_one_batch(model, other_args):
     command = [
         "python3",
         "-m",
-        "sglang.bench_one_batch",
+        "sglang.benchmark.one_batch",
         "--batch-size",
         "1",
         "--input",
